Split mentorController further
