+++ .eslintrc
@@ -6,7 +6,8 @@
   "rules": {
     "indent": [
       "error",
-      4
+      4,
+      { "SwitchCase": 1 }
     ],
     "linebreak-style": [
       0
